console: add line carrying backslash
When using interactive console(stdin) instead of \set delimiter <delimiter>
with "\", "\" in the end if line can be used. Works on both server/client
side.

Local console is impossible to test with a pipe or tatantool interactive
mode. The problem with interactive mode (-i) is that the flag that sets it,
does not influence the isatty() check, while startup script is executed
(#5064). To overload this check and test the local console isatty lib
was added.