replication: track information about replica
This is a draft for the patch.
The patch allows to track information about changing relay state. At
every change of relay state timestamp, vclock, new state (and error
message if exists) will be saved at _cluster space.

The patch adds trigger list at relay, that is invoked when relay changes
its state.  The trigger that updates _cluster space is setted when a
replica is registered.